+++ src/main/java/org/mockito/configuration/AnnotationEngine.java
@@ -10,16 +10,14 @@ import org.mockito.MockitoAnnotations;
  * Configures mock creation logic behind &#064;Mock, &#064;Captor and &#064;Spy annotations
  * <p>
  * If you are interested then see implementations or source code of {@link MockitoAnnotations#initMocks(Object)}
+ *
+ * <p>This interface can be used to configure a different annotation engine through
+ * {@link org.mockito.configuration.IMockitoConfiguration}, however this mechanism is being superseded by the new
+ * {@link org.mockito.plugins plugin} system.
+ *
+ * @deprecated Please use {@link org.mockito.plugins.AnnotationEngine} instead,
+ *             this interface will probably be removed in mockito 3.
  */
-public interface AnnotationEngine {
-
-    /**
-     * Allows extending the interface to perform action on specific fields on the test class.
-     * <p>
-     * See the implementation of this method to figure out what is it for.
-     * 
-     * @param clazz Class where to extract field information, check implementation for details
-     * @param testInstance Test instance
-     */
-    void process(Class<?> clazz, Object testInstance);
+@Deprecated
+public interface AnnotationEngine extends org.mockito.plugins.AnnotationEngine {
 }
