--- src/test/java/org/mockitousage/configuration/SmartMock.java
@@ -1,15 +1,0 @@
-/*
- * Copyright (c) 2007 Mockito contributors
- * This program is made available under the terms of the MIT License.
- */
-package org.mockitousage.configuration;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-import static java.lang.annotation.ElementType.FIELD;
-
-@Target( { FIELD })
-@Retention(RetentionPolicy.RUNTIME)
-public @interface SmartMock {}
